ref: separate notification form CallActivity - create CallNotificationManager to handle notifications, to ensure separation of concerns - create helper class CallContactAvatarHelper to get contact's avatar. It is needed by the CallActivity and CallNotificationManager
